Add drag&drop component and func, add styling
