+++ src/components/media-importer.jsx
@@ -1,85 +1,13 @@
 import React, { useState, useEffect, useRef } from 'react'
 // import { queryClient } from '../media'
 import { useQuery } from 'react-query'
-import { getAllAssets, uploadFile } from '../util/media-importer'
-
-import LoadingIcon from './loading-icon'
+import DragAndDrop from './drag-and-drop'
+import { getAllAssets, uploadFile, loadPickedAsset } from '../util/media-importer'
 import './media.scss'
-
-const DragAndDrop = ({ children }) => {
-	const mount = useRef(false)
-	const [state, setState] = useState(null)
-
-	const handleDragEvent = (ev) => {
-		ev.preventDefault()
-		ev.stopPropagation()
-	}
-
-	const handleOnChange = (ev) => {
-		ev.preventDefault()
-		ev.stopPropagation()
-
-		setState(ev)
-		uploadFile(ev)
-	}
-
-	useEffect(() => {
-		if (mount.current) {
-			console.log(state)
-			// uploadFile(state)
-		}
-	}, [state])
-
-	useEffect(() => {
-		mount.current = true
-		// document.addEventListener('dragenter', handleDragEvent)
-		// document.addEventListener('dragleave', handleDragEvent)
-		// document.addEventListener('drag', handleDragEvent)
-		// document.addEventListener('change', handleOnChange)
-		return () => {
-			mount.current = false
-			// document.removeEventListener('dragenter', handleDragEvent)
-			// document.removeEventListener('dragleave', handleDragEvent)
-			// document.removeEventListener('drag', handleDragEvent)
-			// document.removeEventListener('change', handleOnChange)
-		}
-	}, [])
-
-	return (
-		<div
-			id="drag-and-drop"
-			onDragEnter={(ev) => {
-				handleDragEvent(ev)
-			}}
-			onDragEnd={(ev) => {
-				handleDragEvent(ev)
-			}}
-			onDrag={(ev) => {
-				handleDragEvent(ev)
-			}}
-			onChange={(ev) => {
-				handleOnChange(ev)
-			}}
-		>
-			{children}
-		</div>
-	)
-}
 
 const MediaImporter = () => {
 	const mounted = useRef(false)
-	const [selectedFile, setSelectedFile] = useState('wait')
-	const [sortOrder, setSortOrder] = useState(false)
 	const { data: listOfAssets, isSuccess } = useQuery('assets', getAllAssets)
-
-	const onBrowseClick = (ev) => {
-		console.log(ev)
-		setSelectedFile(ev)
-	}
-
-	const sendMediaToCreator = (media) => {
-		return window.Materia.onMediaImportComplete([media])
-	}
 
 	/**
 	 * It returns a card component that contains the assets data.
@@ -87,7 +15,7 @@ const MediaImporter = () => {
 	 */
 	const AssetCard = ({ name, thumb, created, type, media }) => {
 		return (
-			<div className="file-info" onClick={() => sendMediaToCreator([media])}>
+			<div className="file-info" onClick={() => loadPickedAsset(media)}>
 				<span className="file-thumbnail">
 					<img src={thumb} alt={name} />
 				</span>
@@ -154,10 +82,8 @@ const MediaImporter = () => {
 			<section id="left-pane">
 				<div className="plane-header">Upload a new file</div>
 
-				<DragAndDrop>
-					<div id="drag-wrapper">
-						<div className="drag-text">Drag a file here to upload</div>
-					</div>
+				<DragAndDrop parseMethod={uploadFile} idStr={'drag-wrapper'}>
+					<div className="drag-text">Drag a file here to upload</div>
 				</DragAndDrop>
 				<div className="drag-footer">
 					<label>
